requirements.txt
google-api-python-client==2.116.0
oauth2client==4.1.3
mcrcon==0.6
google-auth
google-auth-oauthlib 
google-auth-httplib2
keyboard
python-dateutil